refactor InputField out of TextField
